Rename "refinement" to "decimation"
