+++ sxs/utilities/decimation/greedy_spline.py
