+++ sxs/utilities/decimation/linear_bisection.py
